% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg_ggsave300dpi.R
\name{eg_ggsave}
\alias{eg_ggsave}
\title{An Example for 'ggplot' with Errorbar}
\usage{
eg_ggsave()
}
\description{
An exmaple.
}
\examples{
library(R11)
library(ggplot2)
x=c(1,3,8)
y=c(5,6,1)
dat=data.frame(x,y)

ggplot(dat, aes(x = x, y = y)) + geom_line() + theme_bw()
p <- last_plot()
ggsave("plot1.pdf", p, width = 3.15, height = 3.15)
ggsave("plot2.tiff", p, width = 4, height = 3, units="cm", dpi=300)
dev.off()
# ggsave(filename, plot = last_plot(), device = NULL, path = NULL,
#       scale = 1, width = NA, height = NA, units = c("in", "cm", "mm"),
#       dpi = 300, limitsize = TRUE, ...)
### 'device' can be c("eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg", "wmf").
###  'units' (of 'width', 'height') can be c("in", "cm", or "mm").
###  'scale' refer to "multiplicative scaling factor".
}
\seealso{
\code{\link[ggplot2]}
}
