% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbar.R
\name{ggbar_MeanSD}
\alias{ggbar_MeanSD}
\title{Bar Charts Showing Mean and SD}
\usage{
ggbar_MeanSD(x, y, fill.bar = NULL, width.bar = 0.5, col.border = NULL,
  size.border = 1, alpha = 1, width.errorbar = 0.3, col.errorbar = NULL,
  size.errorbar = 1)
}
\description{
Dataset contains only 'y' and 'factor(x)'. The heights of the bars represent y-values against x-factors.
}
\examples{
library(R11)
y = iris$Sepal.Length
x = iris$Species
# vertical
ggbar_MeanSD(x, y)
# horizontal
ggbar_MeanSD(x, y, alpha = 0.5) +
  coord_flip()
# set colors
ggbar_MeanSD(x, y, fill.bar = c("grey10", "grey50", "grey70"))+
  theme_bw()
# set ylim using coor_cartesian()
ggbar_MeanSD(x, y, alpha = 0.5) +
   coord_cartesian(ylim = c(0, 10))

}
\seealso{
\code{\link[ggplot2]{geom_bar}}, \code{\link[tidyverse]}
}
