% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggR11.R
\name{theme_ggR11}
\alias{theme_ggR11}
\title{Default theme used in all `R11` package ggplots}
\usage{
theme_ggR11(ggtheme = ggplot2::theme_bw(), ggR11.layer = TRUE)
}
\arguments{
\item{ggtheme}{A function, `ggplot2` theme name. Default value is
`ggplot2::theme_bw()`. Any of the `ggplot2` themes, or themes from
extension packages are allowed (e.g., `ggthemes::theme_economist()`,
`hrbrthemes::theme_ipsum_ps()`, `ggthemes::theme_fivethirtyeight()`).}

\item{ggR11.layer}{Logical that decides whether `theme_ggR11`
theme elements are to be displayed along with the selected `ggtheme`
(Default: `TRUE`).}
}
\value{
A `ggplot2` object with the `theme_ggR11` theme.
}
\description{
Common theme used across all ggplots generated in `R11`
  and *assumed* by the author to be aesthetically pleasing to the
  user/reader.
}
\references{
\url{https://indrajeetpatil.github.io/ggR11/articles/theme_ggR11.html}
}
\author{
Indrajeet Patil, Weiping Mei
}
