% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Data Summary}
\usage{
data_summary(x, digits = 3)
}
\description{
The results include sumple number (N), mean, varicance (Var), standard deviation (sd), standard error (se, =sd/sqrt(n)), coefficient of variation (CV, %),
 median, median absolute deviation (mad), the first quartile (Q1), the third quartile (Q3), minimum, maximum, corrected sum or squares (CSS), uncorrected sum or squares (USS), range, interquartile range (IQR).
}
\examples{
library(R11)
x<-rnorm(n = 15, mean = 100, sd = 10) 
data_summary(x)
}
\seealso{
\code{\link[stats]{IQR}},\code{\link[stats]{mad}}, \code{\link[DescTools]{MeanCI}}, \code{\link[DescTools]{MedianCI}}, \code{\link[DescTools]{MeanCI}}, \code{\link[DescTools]{Skew}}, \code{\link[DescTools]{Kurt}}
}
