% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Data Summary}
\usage{
data_summary(x, digits = 3)
}
\description{
The results include sumple number (N), mean, varicance (Var), standard deviation (sd), standard error (se, =sd/sqrt(n)), coefficient of variation (CV, %),
 median,  the first quartile (Q1), the third quartile (Q3), corrected sum or squares (CSS), uncorrected sum or squares (USS), range, interquartile range (IQR), Skewness, and Kurtosis.
}
\examples{
library(research1p1)
x<-rnorm(n = 15, mean = 100, sd = 10) 
data_summary(x)
}
